Land readers on the dataset, not on its config
`/datasets/<ref>` was the config page — model version, files, GitHub settings.
That is the operator's view, and it was the first thing anyone following a
dataset link saw. Config moves to `/datasets/<ref>/config`, and the root becomes
a redirect to whatever the dataset actually offers:

  1. its first dashboard — which is the About page when the repo ships one,
     since that already leads the list;
  2. AI Q&A, when questions have been asked;
  3. ltool with the first source selected, when there is neither.

The third tier is the one worth stating: a dataset with no dashboards and no
questions is EMPTY, not broken, and the useful thing to hand someone then is a
query surface pointed at its data. ltool with no source is a blank picker — the
same dead end the config page was — so the source is what makes it land
somewhere.

The chain is a pure function so it can be tested without a database; the page
only does the I/O. It is a server component, so the redirect happens before
anything renders, and it stays DuckDB-free like every page.

Also add a "Query" item to the dashboard nav, beside AI Q&A — the same kind of
thing, a way into the data that isn't a dashboard someone built in advance. It
passes the dataset ID (the route param may be a name) and the first source, so
ltool opens on a `run: <source> ->` starter.

Three links moved with the route: the nav's config item, the home page's gear,
and the home page's "+N more" under a question list — that one now goes to the
Q&A page, which is what it always meant.
